requirements.txt
google-generativeai
sklearn
joblib
pandas